split out builder, made knuth.csv
